Flytt rapids-filene inn i riktige pakker
