+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/innboks/InnboksSink.kt
@@ -1,4 +1,4 @@
-package no.nav.personbruker.dittnav.eventaggregator.varsel
+package no.nav.personbruker.dittnav.eventaggregator.innboks
 
 import kotlinx.coroutines.runBlocking
 import no.nav.helse.rapids_rivers.JsonMessage
@@ -7,14 +7,14 @@ import no.nav.helse.rapids_rivers.MessageProblems
 import no.nav.helse.rapids_rivers.RapidsConnection
 import no.nav.helse.rapids_rivers.River
 import no.nav.helse.rapids_rivers.asLocalDateTime
-import no.nav.helse.rapids_rivers.asOptionalLocalDateTime
 import no.nav.personbruker.dittnav.eventaggregator.common.LocalDateTimeHelper.nowAtUtc
 import no.nav.personbruker.dittnav.eventaggregator.config.EventType
-import no.nav.personbruker.dittnav.eventaggregator.oppgave.Oppgave
+import no.nav.personbruker.dittnav.eventaggregator.metrics.RapidMetricsProbe
+import no.nav.personbruker.dittnav.eventaggregator.varsel.VarselRepository
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
-internal class OppgaveSink(
+internal class InnboksSink(
     rapidsConnection: RapidsConnection,
     private val varselRepository: VarselRepository,
     private val rapidMetricsProbe: RapidMetricsProbe,
@@ -22,11 +22,11 @@ internal class OppgaveSink(
 ) :
     River.PacketListener {
 
-    private val log: Logger = LoggerFactory.getLogger(OppgaveSink::class.java)
+    private val log: Logger = LoggerFactory.getLogger(InnboksSink::class.java)
 
     init {
         River(rapidsConnection).apply {
-            validate { it.demandValue("@event_name", "oppgave") }
+            validate { it.demandValue("@event_name", "innboks") }
             validate { it.demandValue("aktiv", true) }
             validate { it.requireKey(
                 "namespace",
@@ -39,12 +39,12 @@ internal class OppgaveSink(
                 "sikkerhetsnivaa",
                 "eksternVarsling"
             ) }
-            validate { it.interestedIn("synligFremTil", "prefererteKanaler")}
+            validate { it.interestedIn("prefererteKanaler")}
         }.register(this)
     }
 
     override fun onPacket(packet: JsonMessage, context: MessageContext) {
-        val oppgave = Oppgave(
+        val innboks = Innboks(
             id = null,
             systembruker = "N/A",
             namespace = packet["namespace"].textValue(),
@@ -58,7 +58,6 @@ internal class OppgaveSink(
             link = packet["link"].textValue(),
             sikkerhetsnivaa = packet["sikkerhetsnivaa"].intValue(),
             sistOppdatert = nowAtUtc(),
-            synligFremTil = packet["synligFremTil"].asOptionalLocalDateTime(),
             aktiv = packet["aktiv"].booleanValue(),
             eksternVarsling = packet["eksternVarsling"].booleanValue(),
             prefererteKanaler = packet["prefererteKanaler"].map { it.textValue() }
@@ -66,12 +65,12 @@ internal class OppgaveSink(
 
         runBlocking {
             if(writeToDb) {
-                varselRepository.persistOppgave(oppgave)
-                log.info("Behandlet oppgave fra rapid med eventid ${oppgave.eventId}")
+                varselRepository.persistVarsel(innboks)
+                log.info("Behandlet innboks fra rapid med eventid ${innboks.eventId}")
             } else {
-                log.info("Dryrun: oppgave fra rapid med eventid ${oppgave.eventId}")
+                log.info("Dryrun: innboks fra rapid med eventid ${innboks.eventId}")
             }
-            rapidMetricsProbe.countProcessed(EventType.OPPGAVE_INTERN, oppgave.appnavn)
+            rapidMetricsProbe.countProcessed(EventType.INNBOKS_INTERN, innboks.appnavn)
         }
     }
